pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-brief"
version = "0.1.0"
description = "Daily threat intelligence briefing generator"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
]

[project.scripts]
threat-brief = "threat_brief.cli:main"

[tool.setuptools.packages.find]
include = ["threat_brief*"]
